varco_fastapi/pyproject.toml
[project]
name          = "varco-fastapi"
version       = "0.0.1"
description   = "FastAPI adapter for varco — routing, auth middleware, job running, typed HTTP client, and DI wiring"
authors       = [{ name = "edoardo.scarpaci", email = "edoardo.scarpaci@gmail.com" }]
readme        = "README.md"
license       = { text = "Apache-2.0" }
requires-python = ">=3.12"
keywords      = ["fastapi", "async", "rest-api", "domain-model", "varco", "middleware", "auth", "jobs"]
classifiers   = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Framework :: FastAPI",
    "Typing :: Typed",
]
dependencies = [
    # Workspace sibling — resolved from the monorepo via [tool.uv.sources] below.
    "varco-core",
    # FastAPI provides the HTTP framework, routing, and OpenAPI schema generation.
    "fastapi>=0.115.0",
    # httpx is the async HTTP client used by AsyncVarcoClient and SyncVarcoClient.
    "httpx>=0.27.0",
    # pydantic-settings for env-var-based configuration (CORSConfig, TrustStore, etc.)
    "pydantic-settings>=2.0",
    # providify for DI wiring (@Configuration, @Provider, Inject[], DIContainer)
    "providify>=0.1.4a3",
    # python-multipart is required for FastAPI form data parsing (file uploads, etc.)
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
# WebSocket and SSE support (requires varco_ws from the workspace)
ws  = ["varco-ws"]
# MCP (Model Context Protocol) adapter support — future
mcp = ["mcp>=1.0"]
# A2A (Agent-to-Agent / Google Agent protocol) support — future
a2a = []
# OpenTelemetry instrumentation — optional; gracefully no-ops when absent
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
]

[project.urls]
Homepage   = "https://github.com/edoardoscarpaci/varco"
Repository = "https://github.com/edoardoscarpaci/varco/tree/main/varco_fastapi"
Issues     = "https://github.com/edoardoscarpaci/varco/issues"

[tool.uv.sources]
# Resolve workspace siblings from the monorepo instead of PyPI.
varco-core = { workspace = true }
varco-ws   = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    # httpx is used as a test client via fastapi.testclient.TestClient + AsyncClient
    "httpx>=0.27.0",
    # uvicorn for integration test server
    "uvicorn>=0.30",
    # starlette test client
    "anyio>=4.0",
    # cryptography for JWT auth tests
    "cryptography>=42.0",
]

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["varco_fastapi"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
markers      = [
    "integration: real external services required (disabled by default — use -m integration)",
]
